start_code/audio_book/requirements.txt
# Core dependencies
pdfplumber>=0.10.0
coqui-tts>=0.27.0
torch>=2.0.0
pydub>=0.25.0
pyyaml>=6.0.0
langdetect>=1.0.9

# LLM API client (DeepSeek compatible with OpenAI API)
openai>=1.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
